+++ app/(app)/(tabs)/explore.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import { Text, View } from "react-native";
+
+export default function explore() {
+  return (
+    <View>
+      <Text>explore</Text>
+    </View>
+  );
+}
