+++ app/(app)/(tabs)/index.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import { Text, View } from "react-native";
+
+export default function Index() {
+  return (
+    <View>
+      <Text>Index</Text>
+    </View>
+  );
+}
