+++ app/(app)/(tabs)/profile.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import { Text, View } from "react-native";
+
+export default function Profile() {
+  return (
+    <View>
+      <Text>Profile</Text>
+    </View>
+  );
+}
