+++ app/(app)/global.css
