--- app/index.tsx
@@ -1,14 +1,0 @@
-import React from 'react'
-import { View ,Text} from 'react-native'
-
-export default function Index() {
-  return (
-    <View style={{
-        flex:1,
-        justifyContent:"center",
-        alignItems:"center",
-    }}>
-        <Text className='font-bold'>Javascript</Text>
-    </View>
-  )
-}
